basic refactor
